tests/requirements-test.txt
# Test dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-timeout==2.1.0
pytest-xdist==3.3.1
pytest-html==3.2.0
pytest-docker==2.0.1
click==8.1.3
rich==13.5.2
pyyaml==6.0
docker==6.1.3
requests==2.31.0
psutil==5.9.5
